--- src/lexer/lex_init.c
@@ -1,20 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   lex_init.c                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: bthomas <bthomas@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/07/08 16:13:21 by bthomas           #+#    #+#             */
-/*   Updated: 2024/07/10 14:14:13 by bthomas          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "lexer.h"
-
-int	init_lex(t_data *data, char **env)
-{
-	data->token = NULL;
-	data->env = env;
-	return (0);
-}
